+++ src/socket.cpp
@@ -18,7 +18,7 @@ using boost::lambda::_2;
 namespace nvim {
 
 void Socket::connect_tcp(const std::string& host, 
-                     const std::string& service, double timeout_sec)
+                     const std::string& service, long timeout_sec)
 {
     tcp::resolver::query query(host, service);
     tcp::resolver::iterator iter = tcp::resolver(io_service_).resolve(query);
@@ -40,7 +40,7 @@ void Socket::connect_tcp(const std::string& host,
         ec ? ec : boost::asio::error::host_not_found);
 }
 
-size_t Socket::read(char *rbuf, size_t capacity, double timeout_sec) {
+size_t Socket::read(char *rbuf, size_t capacity, long timeout_sec) {
     deadline_.expires_from_now(boost::posix_time::seconds(timeout_sec));
     boost::system::error_code ec = boost::asio::error::would_block;
     size_t rlen;
@@ -54,7 +54,7 @@ size_t Socket::read(char *rbuf, size_t capacity, double timeout_sec) {
     return rlen;
 }
 
-void Socket::write(char *sbuf, size_t size, double timeout_sec) {
+void Socket::write(char *sbuf, size_t size, long timeout_sec) {
     deadline_.expires_from_now(boost::posix_time::seconds(timeout_sec));
     boost::system::error_code ec = boost::asio::error::would_block;
     boost::asio::async_write(socket_, boost::asio::buffer(sbuf, size), var(ec) = _1);
